add ag grid for search results
