+++ frontend/src/types/Search.ts
@@ -1,0 +1,11 @@
+export interface SearchBarProps {
+  onSearch?: (filters: SearchFilters) => void;
+}
+
+export interface SearchFilters {
+  query: string;
+  description: string;
+  actor: string;
+  year: string;
+  minRating: string;
+}
